AbilityProperty and MovingProperty introduced and started working on Knight implementation actually.
The actual framework of Unit is like 80% complete. It's concepts have
been thought through fully, I just have to reorganize and make sure all
ideas are sound before heavily implementing them. I'm pretty sure this
is the right approach to using AbilityProperties.
Started working on knight implementation actually and should be ready to
work on all the rest of the units completely in the next couple of
commits.